fix(assessments): FM-021 64-bit skip multiply + drop dead using (fresh-review)
Fresh adversarial review = SAFE TO SHIP. Folds its 1 LOW: (page-1)*limit was
int32 and could overflow to a small OFFSET for a pathological admin page value,
silently serving wrong rows (worse than legacy's 500). Now computed in long
(OFFSET is bigint); pinned by a golden test. Also removes a dead
System.Globalization using left after the shared-mapper extraction.

dotnet 0/0; 184u+1c+123i.

+++ services/api/src/FormMaps.Application/Assessments/IAllResultsReader.cs
@@ -11,5 +11,5 @@ public sealed record AllResultsPage(IReadOnlyList<PcaHistorySession> Rows, int T
 /// </summary>
 public interface IAllResultsReader
 {
-    Task<AllResultsPage> ReadAsync(RequestContext context, int skip, int limit, CancellationToken cancellationToken = default);
+    Task<AllResultsPage> ReadAsync(RequestContext context, long skip, int limit, CancellationToken cancellationToken = default);
 }
